+++ src/packages/babel-preset-kiwicom/src/__tests__/transformations.test.js
@@ -31,3 +31,4 @@ function transform(target) {
 
 generateTestsFromFixtures(`${__dirname}/__fixtures__`, transform('js'));
 generateTestsFromFixtures(`${__dirname}/__fixtures__`, transform('flow'));
+generateTestsFromFixtures(`${__dirname}/__fixtures__`, transform('js-esm'));